majit: share bridge-entry virtual materialization via resume_box_reader
Move the trace-emit (box-reader) flavour of the resume-data reader out of
pyre-jit-trace/state.rs into a new majit-metainterp/resume_box_reader.rs so
every #[jit_interp] consumer can materialize a guard's virtuals and replay its
deferred heap writes at bridge entry, not only the hand-written PyreJitState.

- Add majit-metainterp/src/resume_box_reader.rs (BridgeVirtualCache,
  materialize_bridge_virtual, replay_pending_fields, emit_pending_field_op,
  decode_fieldnum, rebuilt_value_to_opref, default_bridge_array_descr) and
  re-export it from majit-metainterp/src/lib.rs.
- pyre-jit-trace/src/state.rs: drop the local BridgeVirtualCache /
  materialize_virtual_from_rd / emit_stroruni_oopspec_call definitions and use
  the shared ones.
- jit_interp macro setup_bridge_sym (codegen_state.rs): call
  replay_pending_fields once, and materialize_bridge_virtual for
  RebuiltValue::Virtual int/ref scalars that previously fell through unseeded
  (bridge read size>chain and dereferenced a NULL node head). Add env-gated
  (AHEUI_BRIDGE_DIAG) bridge-seed diagnostics.
- Hoist field_descr_from_parent_by_offset into majit-ir/src/descr.rs;
  pyre-jit-trace make_field_descr_with_parent delegates to it.
