+++ majit/majit-metainterp/src/lib.rs
@@ -93,6 +93,7 @@ pub mod quasiimmut;
 pub mod recorder;
 pub mod resoperation;
 pub mod resume;
+pub mod resume_box_reader;
 pub(crate) mod ruleopt;
 pub mod support;
 mod trace_ctx;
@@ -154,6 +155,10 @@ pub use pyjitpl::{
     trace_jitcode_from_merge_point, trace_jitcode_with_args, trace_jitcode_with_args_and_runtime,
 };
 pub use quasiimmut::QuasiImmut;
+pub use resume_box_reader::{
+    BridgeVirtualCache, decode_fieldnum, default_bridge_array_descr, emit_pending_field_op,
+    materialize_bridge_virtual, rebuilt_value_to_opref, replay_pending_fields,
+};
 pub use trace_ctx::BridgeInlineCarrier;
 pub use trace_ctx::GreenBox;
 pub use trace_ctx::MergePoint;
